+++ R/RCM2MOM.R
@@ -51,9 +51,9 @@ RCM2MOM <- function(RCModel) {
       aperm(c(2, 1, 3))
     
     if (!is.null(cp$SLarray)) {
-      cp$SLarray <- lapply(report, make_SL, sel_block = RCModel@data@sel_block) %>% 
+      cp$SLarray <- lapply(report, make_SL, sel_block = RCModel@data@sel_block, f = f) %>% 
         lapply(expand_V_matrix, nyears = RCModel@OM@nyears, proyears = RCModel@OM@proyears) %>% 
-        simplify2array() %>% aperm(c(3, 1, 2))
+        simplify2array() %>% aperm(3:1)
     }
     
     if (!is.null(cp$Data)) {